requirements.txt
appdirs==1.4.4
asgiref==3.3.4
attrs==21.2.0
backcall==0.2.0
black==20.8b1
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
coreapi==2.3.3
coreschema==0.0.4
coverage==5.3
decorator==5.0.9
Django==3.2.4
django-cors-headers==3.7.0
django-coverage-plugin==1.8.0
django-debug-toolbar==3.1.1
django-extensions==3.0.9
django-rest-framework==0.1.0
django-stubs==1.5.0
djangorestframework==3.12.4
drf-yasg==1.20.0
factory-boy==3.2.0
Faker==8.8.0
flake8==3.8.3
flake8-isort==4.0.0
idna==2.10
inflection==0.5.1
iniconfig==1.1.1
ipdb==0.13.3
ipython==7.24.1
ipython-genutils==0.2.0
isort==5.5.2
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mccabe==0.6.1
more-itertools==8.8.0
mypy==0.770
mypy-extensions==0.4.3
packaging==20.9
parso==0.8.2
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.18
ptyprocess==0.7.0
py==1.10.0
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.9.0
pyparsing==2.4.7
pytest==6.0.1
pytest-asyncio==0.14.0
pytest-cov==2.10.0
pytest-django==3.9.0
pytest-sugar==0.9.4
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
ruamel.yaml==0.17.9
ruamel.yaml.clib==0.2.2
six==1.16.0
sqlparse==0.4.1
termcolor==1.1.0
testfixtures==6.17.1
text-unidecode==1.3
toml==0.10.2
traitlets==5.0.5
typed-ast==1.4.3
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.5
wcwidth==0.2.5
Werkzeug==1.0.1
